inandout: Move configurator implementation in own class
This makes the factory implementation leaner and separates the
configurator implementation in an own class.
